+++ components/alibi-detect-server/Dockerfile
@@ -1,4 +1,4 @@
-# TODO: Add to release script 
+# TODO: Add to release script
 FROM docker.io/seldonio/seldon-core-s2i-python37-ubi8:1.8.0-dev
 LABEL name="Seldon Alibi Detect Server" \
       vendor="Seldon Technologies" \
@@ -6,6 +6,13 @@ LABEL name="Seldon Alibi Detect Server" \
       release="1" \
       summary="Alibi Detect Server for Seldon Core" \
       description="The Alibi Detect Server provides outlier, drift and adversarial detection services for Seldon Core"
+
+# Install Rclone Binary to be present in the image
+RUN yum install -y unzip
+RUN wget https://downloads.rclone.org/v1.55.1/rclone-v1.55.1-linux-amd64.zip && \
+    unzip rclone-v1.55.1-linux-amd64.zip && \
+    mv rclone-v1.55.1-linux-amd64/rclone /usr/bin/rclone && \
+    rm -rf rclone-v1.55.1-linux-amd64.zip rclone-v1.55.1-linux-amd64
 
 ADD requirements_server.txt .
