+++ components/alibi-detect-server/adserver/base/storage.py
@@ -1,7 +1,19 @@
 import os
 import kfserving
+import logging
+import tempfile
 from distutils.util import strtobool
-from seldon_core.storage import Rclone
+
+
+logger = logging.getLogger(__name__)
+
+try:
+    from sh import rclone
+except ImportError:
+    logger.warning(
+        "rclone-based storage funcionality not available without rclone binary"
+    )
+    rclone = None
 
 
 def _rclone_enabled():
@@ -19,6 +31,27 @@ def _rclone_enabled():
 
 
 RCLONE_ENABLED = _rclone_enabled()
+
+
+class Rclone:
+    def __init__(self, cfg_file: str = None):
+        self.cfg_file = cfg_file
+
+    def copy(self, src: str, dest: str = None):
+        if rclone is None:
+            raise RuntimeError(
+                "rclone binary not found - rclone-based storage funcionality disabled"
+            )
+
+        if dest is None:
+            dest = tempfile.mkdtemp()
+
+        kwargs = {}
+        if self.cfg_file is not None:
+            kwargs["config"] = os.path.abspath(os.path.expanduser(self.cfg_file))
+
+        rclone.copy(src, dest, **kwargs)
+        return dest
 
 
 def download_model(storage_uri) -> str:
