+++ python/seldon_core/storage.py
@@ -24,37 +24,6 @@ import tempfile
 
 _LOCAL_PREFIX = "file://"
 
-logger = logging.getLogger(__name__)
-
-try:
-    from sh import rclone
-except ImportError:
-    logger.warning(
-        "rclone-based storage funcionality not available without rclone binary"
-    )
-    rclone = None
-
-
-class Rclone:
-    def __init__(self, cfg_file: str = None):
-        self.cfg_file = cfg_file
-
-    def copy(self, src: str, dest: str = None):
-        if rclone is None:
-            raise RuntimeError(
-                "rclone binary not found - rclone-based storage funcionalty disabled"
-            )
-
-        if dest is None:
-            dest = tempfile.mkdtemp()
-
-        kwargs = {}
-        if self.cfg_file is not None:
-            kwargs["config"] = os.path.abspath(os.path.expanduser(self.cfg_file))
-
-        rclone.copy(src, dest, **kwargs)
-        return dest
-
 
 class Storage:
     @staticmethod
